pyproject.toml
[tool.poetry]
name = "drf-api-2"
version = "0.1.0"
description = ""
authors = ["Mahmoud alzoubi <mahmoudalzoubi073@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
